+++ OVRLighthouseManager/Contracts/Services/IMiscSettingsService.cs
@@ -7,11 +7,17 @@ using System.Threading.Tasks;
 namespace OVRLighthouseManager.Contracts.Services;
 public interface IMiscSettingsService
 {
+    public bool MinimizeOnLaunchedByOpenVR
+    {
+        get;
+    }
+
     public bool OutputDebug
     {
         get;
     }
 
     public Task InitializeAsync();
+    public Task SetMinimizeOnLaunchedByOpenVRAsync(bool minimizeOnLaunchedByOpenVR);
     public Task SetOutputDebugAsync(bool outputDebug);
 }