+++ OVRLighthouseManager/Services/MiscSettingsService.cs
@@ -3,7 +3,13 @@
 namespace OVRLighthouseManager.Services;
 internal class MiscSettingsService : IMiscSettingsService
 {
+    private const string SettingsKey_MinimizeOnLaunchedByOpenVR = "MinimizeOnLaunchedByOpenVR";
     private const string SettingsKey_OutputDebug = "OutputDebug";
+
+    public bool MinimizeOnLaunchedByOpenVR
+    {
+        get; set;
+    }
 
     public bool OutputDebug
     {
@@ -17,6 +23,12 @@ internal class MiscSettingsService : IMiscSettingsService
         _localSettingsService = localSettingsService;
     }
 
+    public async Task SetMinimizeOnLaunchedByOpenVRAsync(bool minimizeOnLaunchedByOpenVR)
+    {
+        MinimizeOnLaunchedByOpenVR = minimizeOnLaunchedByOpenVR;
+        await SaveMinimizeOnLaunchedByOpenVRInSettingsAsync(MinimizeOnLaunchedByOpenVR);
+    }
+
     public async Task SetOutputDebugAsync(bool outputDebug)
     {
         OutputDebug = outputDebug;
@@ -26,7 +38,18 @@ internal class MiscSettingsService : IMiscSettingsService
     public async Task InitializeAsync()
     {
         OutputDebug = await LoadOutputDebugFromSettingsAsync();
+        MinimizeOnLaunchedByOpenVR = await LoadMinimizeOnLaunchedByOpenVRFromSettingsAsync();
         await Task.CompletedTask;
+    }
+
+    private async Task SaveMinimizeOnLaunchedByOpenVRInSettingsAsync(bool minimizeOnLaunchedByOpenVR)
+    {
+        await _localSettingsService.SaveSettingAsync(SettingsKey_MinimizeOnLaunchedByOpenVR, minimizeOnLaunchedByOpenVR);
+    }
+
+    private async Task<bool> LoadMinimizeOnLaunchedByOpenVRFromSettingsAsync()
+    {
+        return await _localSettingsService.ReadSettingAsync<bool>(SettingsKey_MinimizeOnLaunchedByOpenVR);
     }
 
     private async Task SaveOutputDebugInSettingsAsync(bool outputDebug)